Don't open quickfix after synchronous dispatch
If another plugin has wrapped :Make in a construct that changes and
restores the current directory, opening the quickfix after a synchronous
make means restoring the directory for the wrong window.  Running
synchronously is already a vastly different experience than the other
adapters, so I think an additional small inconsistency is acceptable
price to pay for rectifying this behavior.