Fixed #108. For using windows gvim with cygwin shells
The actually problem is that vim automatically set &shellquote='"' when
&shell is some kind of posix-like shells. When &shellquote is '"', writing
double quotes in the command passed to system() would cause the command to
be misinterpreted, and producing the weird error as reported in #108.
The solution is to detect whether shellquote is set to double quote, and
alter the sent command accordingly.

Note that this issue would not affect default windows cmd shell, thus the
command should only be altered when *sh shells are used.