Expand `=expr` in arguments
The primary use case here is to set a dynamic default in b:dispatch,
based on the current line number or some other criteria.  :FocusDispatch
has changed to show the raw backtick expressions without a count and the
expansions if a count is given.

In addition to `=expr`, there is also `+=expr`, which expands only with
a line number, and `-=expr`, which expands only without one.  In all
three, v:lnum is set to the line number given or zero.