Add User#DispatchQuickfix, set t:dispatch_quickfix_winnr and break
In `s:open_quickfix` a user autocommand is triggered after the quickfix
window has been massaged, which includes setting `w:quickfix_title`.
This has to be done in the end, since it is not available with `au
FileType qf`.

The variable `t:dispatch_quickfix_winnr` is set to inform the user about the
window's number.