Allow forcing a :Make subcommand to use :Start
Say makeprg is lein, and you want to override `:Make repl` to call not
`:Dispatch lein repl` but rather `:Start lein repl`.  This change
enables achieving that with a special declaration in errorformat:

    setl errorformat+=%\\&terminal=%\\Crepl%\\>%.%#

The pattern uses Vim's peculiar scanf variant documented under the
errorformat help topic.  Note the use of %\C, because Vim is case
insensitive by default in this context.  The key name "terminal" is
meant to convey the underlying semantic: the command requires a
terminal, and should not be piped to a file.

I am experimentally providing this to help with "do what I mean", but it
goes against the spirit of :Make as a paper thin :make wrapper, so even
if it sticks around, I'll still be favoring dedicated commands like
:Console as the official interface.