Set l# alongside v:lnum
This enables line number handling in a way that, unlike `=...`
expressions, is compatible with plain Vim.

    let b:dispatch = 'rspec %:s/$/\=exists("l#") ? ":".l# : "")/'

Assigning to a local variable with a hash symbol in the name isn't
normally allowed, so this is great for avoiding false positives as well
as style.  Older versions of Vim don't accept such variable names with
the l: prefix, so make sure you omit it.

I've also defined a backup variable in case further issues crop up with
the cool, cool name.