Unify expansion
I was going to a great deal of trouble to apply `=...` expansion
separately for 'makeprg' so that QuickFixCmdPre wouldn't get an
inaccurate picture of the command to run.  But I've decided that this
expansion is on the way out and it doesn't really matter if the
auto-command sees some weird syntax.

This also drops some support for experimental variations of that syntax
that were never officially supported.