Report command status in the quickfix title
For example, instead of (tmux/12345), show (12345/Success), or
(12345/Running) if the command hasn't finished yet.

This is helpful in cases where it is difficult at a glance for a user to
tell whether their command succeeded just from reading the output.

This is updated in all cases where a postfix is added to command in
messages to the user, including the quickfix title.

For updating quickfix lists, the title up to the '/' character in the
postfix is used.